+++ api/api_storage.go
@@ -186,7 +186,7 @@ type StorageMiner interface {
 	//
 	// If allowFetch is set, list of paths to which the sector can be fetched will also be returned.
 	// - Paths which have sector files locally (don't require fetching) will be listed first.
-	// - Paths which have sector files locally will not be filtered based on based on AllowTypes/DenyTypes.
+	// - Paths which have sector files locally will not be filtered based on AllowTypes/DenyTypes.
 	// - Paths which require fetching will be filtered based on AllowTypes/DenyTypes. If multiple
 	//   file types are specified, each type will be considered individually, and a union of all paths
 	//   which can accommodate each file type will be returned.